barrouter/requirements.txt
pysqlite==1.0.1
Django==1.3
django-piston==0.2.3
requests==0.6.1
